Fix uuids
