+++ api/stately/loadouts-queries.ts
@@ -23,7 +23,15 @@ import {
   LoadoutShare as StatelyLoadoutShare,
   StatConstraint as StatelyStatConstraint,
 } from './generated/index.js';
-import { batches, listToMap, stripDefaults, stripTypeName, Transaction } from './stately-utils.js';
+import {
+  batches,
+  listToMap,
+  parseUUID,
+  stringifyUUID,
+  stripDefaults,
+  stripTypeName,
+  Transaction,
+} from './stately-utils.js';
 
 export function keyFor(
   platformMembershipId: string | bigint,
@@ -71,34 +79,6 @@ async function getAllLoadoutsForUser(platformMembershipId: string): Promise<
         loadout: a,
       })),
     );
-}
-
-/**
- * Convert array of 16 byte values to UUID string format of the form:
- * XXXXXXXX-XXXX-XXXX-XXXX-XXXXXXXXXXXX
- */
-const byteToHex: string[] = [];
-
-for (let i = 0; i < 256; ++i) {
-  byteToHex.push((i + 0x100).toString(16).slice(1));
-}
-
-// Copied from uuid package but without validation
-function stringifyUUID(arr: Uint8Array, offset = 0) {
-  // Note: Be careful editing this code!  It's been tuned for performance
-  // and works in ways you may not expect. See https://github.com/uuidjs/uuid/pull/434
-  return `${
-    byteToHex[arr[offset + 0]] +
-    byteToHex[arr[offset + 1]] +
-    byteToHex[arr[offset + 2]] +
-    byteToHex[arr[offset + 3]]
-  }-${byteToHex[arr[offset + 4]]}${byteToHex[arr[offset + 5]]}-${byteToHex[arr[offset + 6]]}${
-    byteToHex[arr[offset + 7]]
-  }-${byteToHex[arr[offset + 8]]}${byteToHex[arr[offset + 9]]}-${byteToHex[arr[offset + 10]]}${
-    byteToHex[arr[offset + 11]]
-  }${byteToHex[arr[offset + 12]]}${byteToHex[arr[offset + 13]]}${
-    byteToHex[arr[offset + 14]]
-  }${byteToHex[arr[offset + 15]]}`;
 }
 
 export function convertLoadoutFromStately(item: StatelyLoadout | StatelyLoadoutShare): Loadout {
@@ -204,37 +184,6 @@ function convertLoadoutItemFromStately(item: StatelyLoadoutItem): LoadoutItem {
     result.craftedDate = Number(item.craftedDate);
   }
   return result;
-}
-
-// This is a copy of the UUID parsing code from the uuid package, but without
-// the validation - I don't really care whether it's a perfectly valid UUID,
-// just that it's 16 bytes.
-function parseUUID(uuid: string): Uint8Array {
-  let v;
-  const arr = new Uint8Array(16); // Parse ########-....-....-....-............
-
-  arr[0] = (v = parseInt(uuid.slice(0, 8), 16)) >>> 24;
-  arr[1] = (v >>> 16) & 0xff;
-  arr[2] = (v >>> 8) & 0xff;
-  arr[3] = v & 0xff; // Parse ........-####-....-....-............
-
-  arr[4] = (v = parseInt(uuid.slice(9, 13), 16)) >>> 8;
-  arr[5] = v & 0xff; // Parse ........-....-####-....-............
-
-  arr[6] = (v = parseInt(uuid.slice(14, 18), 16)) >>> 8;
-  arr[7] = v & 0xff; // Parse ........-....-....-####-............
-
-  arr[8] = (v = parseInt(uuid.slice(19, 23), 16)) >>> 8;
-  arr[9] = v & 0xff; // Parse ........-....-....-....-############
-  // (Use "/" to avoid 32-bit truncation when bit-shifting high-order bytes)
-
-  arr[10] = ((v = parseInt(uuid.slice(24, 36), 16)) / 0x10000000000) & 0xff;
-  arr[11] = (v / 0x100000000) & 0xff;
-  arr[12] = (v >>> 24) & 0xff;
-  arr[13] = (v >>> 16) & 0xff;
-  arr[14] = (v >>> 8) & 0xff;
-  arr[15] = v & 0xff;
-  return arr;
 }
 
 export function convertLoadoutToStately(
